+++ src/lib/card/beta/rimland_nomads.rs
@@ -59,7 +59,7 @@ impl Card for RimlandNomads {
         Some(&mut self.unit_base)
     }
 
-    fn on_take_damage(&mut self, state: &State, from: &uuid::Uuid, damage: u8) -> anyhow::Result<Vec<Effect>> {
+    fn on_take_damage(&mut self, state: &State, from: &uuid::Uuid, damage: u16) -> anyhow::Result<Vec<Effect>> {
         let dealer = state.get_card(from);
         let dealer_is_desert = dealer
             .get_site_base()
